home 라우트 생성, 메인메뉴 정의
